requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
autopep8==1.6.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20241021
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-crispy-forms==2.0
django-import-export==4.1.1
django-jazzmin==2.6.2
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-taggit==4.0.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
ecdsa==0.19.0
executing==2.0.1
facebook-sdk==3.1.0
Flask==3.0.3
fonttools==4.51.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.1
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==1.7.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
pywin32==306
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.5.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.40.0
tablib==3.5.0
toml==0.10.2
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.3
gunicorn==23.0.0
whitenoise==6.8.2